backend/requirements.txt
asgiref==3.6.0
pytz==2023.3
sqlparse==0.4.4
Django==3.2
typing_extensions==4.5.0
psycopg2-binary==2.8.6
djangorestframework==3.12.4
requests==2.30.0
PyJWT==2.7.0
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
